Split benchmarks from unit tests
